+++ api/RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Prefer v3 instance types over v2
+
 ## 0.13.14 (2025-10-03)
 
 * Fix choosing appropriate optimal instance instead of expensive GPU instance by default